+++ doomclassic/doom/r_segs.cpp
@@ -368,7 +368,7 @@ R_StoreWallRange
     
     // calculate ::g->rw_distance for scale calculation
     ::g->rw_normalangle = ::g->curline->angle + ANG90;
-	offsetangle = abs((long)(::g->rw_normalangle-::g->rw_angle1));
+	offsetangle = abs((int)(::g->rw_normalangle-::g->rw_angle1));
     
     if (offsetangle > ANG90)
 	offsetangle = ANG90;
