+++ neo/framework/File_SaveGame.h
@@ -2,9 +2,9 @@
 ===========================================================================
 
 Doom 3 BFG Edition GPL Source Code
-Copyright (C) 1993-2012 id Software LLC, a ZeniMax Media company. 
+Copyright (C) 1993-2012 id Software LLC, a ZeniMax Media company.
 
-This file is part of the Doom 3 BFG Edition GPL Source Code ("Doom 3 BFG Edition Source Code").  
+This file is part of the Doom 3 BFG Edition GPL Source Code ("Doom 3 BFG Edition Source Code").
 
 Doom 3 BFG Edition Source Code is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -28,10 +28,11 @@ If you have questions concerning this license or the applicable additional terms
 #ifndef __FILE_SAVEGAME_H__
 #define __FILE_SAVEGAME_H__
 
-#include "zlib/zlib.h"
+#include <zlib.h>
 
 // Listing of the types of files within a savegame package
-enum saveGameType_t {
+enum saveGameType_t
+{
 	SAVEGAMEFILE_NONE			= 0,
 	SAVEGAMEFILE_TEXT			= BIT( 0 ),	// implies that no checksum will be used
 	SAVEGAMEFILE_BINARY			= BIT( 1 ),	// implies that a checksum will also be used
@@ -45,24 +46,28 @@ enum saveGameType_t {
 
 /*
 ================================================
-idFile_SaveGame 
+idFile_SaveGame
 ================================================
 */
-class idFile_SaveGame : public idFile_Memory {
+class idFile_SaveGame : public idFile_Memory
+{
 public:
 	idFile_SaveGame() : type( SAVEGAMEFILE_NONE ), error( false ) {}
-	idFile_SaveGame( const char * _name ) : idFile_Memory( _name ), type( SAVEGAMEFILE_NONE ), error( false ) {}
-	idFile_SaveGame( const char * _name, int type_ ) : idFile_Memory( _name ), type( type_ ), error( false ) {}
+	idFile_SaveGame( const char* _name ) : idFile_Memory( _name ), type( SAVEGAMEFILE_NONE ), error( false ) {}
+	idFile_SaveGame( const char* _name, int type_ ) : idFile_Memory( _name ), type( type_ ), error( false ) {}
 
 	virtual ~idFile_SaveGame() { }
 
-	bool operator==( const idFile_SaveGame & other ) const {
+	bool operator==( const idFile_SaveGame& other ) const
+	{
 		return idStr::Icmp( GetName(), other.GetName() ) == 0;
 	}
-	bool operator==( const char * _name ) const {
+	bool operator==( const char* _name ) const
+	{
 		return idStr::Icmp( GetName(), _name ) == 0;
 	}
-	void SetNameAndType( const char *_name, int _type ) {
+	void SetNameAndType( const char* _name, int _type )
+	{
 		name = _name;
 		type = _type;
 	}
@@ -82,12 +87,14 @@ class idSGFwriteThread;
 class idSGFdecompressThread;
 class idSGFcompressThread;
 
-struct blockForIO_t {
-	byte *		data;
+struct blockForIO_t
+{
+	byte* 		data;
 	size_t		bytes;
 };
 
-class idFile_SaveGamePipelined : public idFile {
+class idFile_SaveGamePipelined : public idFile
+{
 public:
 	// The buffers each hold two blocks of data, so one block can be operated on by
 	// the next part of the generate / compress / IO pipeline.  The factor of two
@@ -97,14 +104,14 @@ public:
 	static const int UNCOMPRESSED_BLOCK_SIZE	= 256 * 1024;
 
 
-							idFile_SaveGamePipelined();
+	idFile_SaveGamePipelined();
 	virtual					~idFile_SaveGamePipelined();
 
-	bool					OpenForReading( const char * const filename, bool useNativeFile );
-	bool					OpenForWriting( const char * const filename, bool useNativeFile );
+	bool					OpenForReading( const char* const filename, bool useNativeFile );
+	bool					OpenForWriting( const char* const filename, bool useNativeFile );
 
-	bool					OpenForReading( idFile * file );
-	bool					OpenForWriting( idFile * file );
+	bool					OpenForReading( idFile* file );
+	bool					OpenForWriting( idFile* file );
 
 	// Finish any reading or writing.
 	void					Finish();
@@ -113,58 +120,96 @@ public:
 	void					Abort();
 
 	// Cancel any reading or writing for app termination
-	static void				CancelToTerminate() { cancelToTerminate = true; }
+	static void				CancelToTerminate()
+	{
+		cancelToTerminate = true;
+	}
 
 	bool					ReadBuildVersion();
-	const char *			GetBuildVersion() const { return buildVersion; }
-	
+	const char* 			GetBuildVersion() const
+	{
+		return buildVersion;
+	}
+
 	bool					ReadSaveFormatVersion();
-	int						GetSaveFormatVersion() const { return saveFormatVersion; }
+	int						GetSaveFormatVersion() const
+	{
+		return saveFormatVersion;
+	}
 	int						GetPointerSize() const;
 
 	//------------------------
 	// idFile Interface
 	//------------------------
 
-	virtual const char *	GetName() const { return name.c_str(); }
-	virtual const char *	GetFullPath() const	{ return name.c_str(); }
-	virtual int				Read( void * buffer, int len );
-	virtual int				Write( const void * buffer, int len );
+	virtual const char* 	GetName() const
+	{
+		return name.c_str();
+	}
+	virtual const char* 	GetFullPath() const
+	{
+		return name.c_str();
+	}
+	virtual int				Read( void* buffer, int len );
+	virtual int				Write( const void* buffer, int len );
 
 	// this file is strictly streaming, you can't seek at all
-	virtual int				Length() const  { return compressedLength; }
-	virtual void			SetLength( size_t len ) { compressedLength = len; }
-	virtual int				Tell() const { assert( 0 ); return 0; }
-	virtual int				Seek( long offset, fsOrigin_t origin ) { assert( 0 ); return 0; }
+	virtual int				Length() const
+	{
+		// RB: 64 bit fix, we don't need support for files bigger than 2 GB
+		return ( int ) compressedLength;
+		// RB end
+	}
+	virtual void			SetLength( size_t len )
+	{
+		compressedLength = len;
+	}
+	virtual int				Tell() const
+	{
+		assert( 0 );
+		return 0;
+	}
+	virtual int				Seek( long offset, fsOrigin_t origin )
+	{
+		assert( 0 );
+		return 0;
+	}
 
-	virtual ID_TIME_T		Timestamp()	const { return 0; }
+	virtual ID_TIME_T		Timestamp()	const
+	{
+		return 0;
+	}
 
 	//------------------------
 	// These can be used by a background thread to read/write data
 	// when the file was opened with 'useNativeFile' set to false.
 	//------------------------
 
-	enum mode_t {
+	enum mode_t
+	{
 		CLOSED,
 		WRITE,
 		READ
 	};
 
 	// Get the file mode: read/write.
-	mode_t					GetMode() const { return mode; }
+	mode_t					GetMode() const
+	{
+		return mode;
+	}
 
 	// Called by a background thread to get the next block to be written out.
 	// This may block until a block has been made available through the pipeline.
 	// Pass in NULL to notify the last write failed.
 	// Returns false if there are no more blocks.
-	bool					NextWriteBlock( blockForIO_t * block );
+	bool					NextWriteBlock( blockForIO_t* block );
 
 	// Called by a background thread to get the next block to read data into and to
 	// report the number of bytes written to the previous block.
 	// This may block until space is available to place the next block.
 	// Pass in NULL to notify the end of the file was reached.
 	// Returns false if there are no more blocks.
-	bool					NextReadBlock( blockForIO_t * block, size_t lastReadBytes );
+	bool					NextReadBlock( blockForIO_t* block, size_t lastReadBytes );
 
 private:
 	friend class idSGFreadThread;
@@ -192,10 +237,10 @@ private:
 	// These variables are used to pass data between threads in a thread-safe manner.
 	//------------------------
 
-	byte *					dataZlib;
+	byte* 					dataZlib;
 	size_t					bytesZlib;
 
-	byte *					dataIO;
+	byte* 					dataIO;
 	size_t					bytesIO;
 
 	//------------------------
@@ -211,7 +256,7 @@ private:
 	// These variables are used by WriteBlock() and ReadBlock().
 	//------------------------
 
-	idFile *				nativeFile;
+	idFile* 				nativeFile;
 	bool					nativeFileEndHit;
 	bool					finished;
 
@@ -219,11 +264,11 @@ private:
 	// The background threads and signals for NextWriteBlock() and NextReadBlock().
 	//------------------------
 
-	idSGFreadThread *		readThread;
-	idSGFwriteThread *		writeThread;
+	idSGFreadThread* 		readThread;
+	idSGFwriteThread* 		writeThread;
 
-	idSGFdecompressThread *	decompressThread;
-	idSGFcompressThread *	compressThread;
+	idSGFdecompressThread* 	decompressThread;
+	idSGFcompressThread* 	compressThread;
 
 	idSysSignal				blockRequested;
 	idSysSignal				blockAvailable;
